+++ scripts/settings.js
@@ -1,12 +1,4 @@
 export const registerSettings = function() {
-  game.settings.register("blind-roll-skills", "forceBlindRolls", {
-    name: game.i18n.localize("BLINDROLLSKILLS.SettingForceBlindRolls"),
-    hint: game.i18n.localize("BLINDROLLSKILLS.SettingForceBlindRollsHint"),
-    scope: "world",
-    config: true,
-    default: true,
-    type: Boolean
-  });
 
   game.settings.register("blind-roll-skills", "showHelpCards", {
     name: game.i18n.localize("BLINDROLLSKILLS.SettingShowHelpCards"),
@@ -16,6 +8,25 @@ export const registerSettings = function() {
     default: true,
     type: Boolean
   });
+
+  // TODO: Implement once modules adopt functionality
+  // game.settings.register("blind-roll-skills", "hideAttacks", {
+  //   name: game.i18n.localize("BLINDROLLSKILLS.SettingHideAttackRolls"),
+  //   scope: "world",
+  //   config: true,
+  //   default: false,
+  //   type: Boolean
+  // });
+
+  //TODO: Implement once modules adopt functionality
+  // game.settings.register("blind-roll-skills", "hideSaves", {
+  //   name: game.i18n.localize("BLINDROLLSKILLS.SettingHideSaves"),
+  //   hint: game.i18n.localize("BLINDROLLSKILLS.SettingHideSavesHint"),
+  //   scope: "world",
+  //   config: true,
+  //   default: false,
+  //   type: Boolean
+  // });
 
   game.settings.register("blind-roll-skills", "hideDeathSaves", {
     name: game.i18n.localize("BLINDROLLSKILLS.SettingHideDeathSaves"),